refactor: Move OpenSearchBackupRepositoryTest to webapp-backups
relates #24869
