+++ webapps-backup/src/test/java/io/camunda/webapps/backup/repository/elasticsearch/TestSnapshotProvider.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.webapps.backup.elasticsearch;
+package io.camunda.webapps.backup.repository.elasticsearch;
 
 import io.camunda.webapps.backup.Metadata;
 import io.camunda.webapps.backup.repository.SnapshotNameProvider;
@@ -24,12 +24,12 @@ public class TestSnapshotProvider implements SnapshotNameProvider {
   }
 
   @Override
-  public Long extractBackupId(final String snapshotName) {
-    return Long.valueOf(snapshotName.split("_")[2]);
+  public String snapshotNamePrefix() {
+    return "test_snapshot_";
   }
 
   @Override
-  public String snapshotNamePrefix() {
-    return "test_snapshot_";
+  public Long extractBackupId(final String snapshotName) {
+    return Long.valueOf(snapshotName.split("_")[2]);
   }
 }
